refactor: move soqlComments.ts and soqlComments.test.ts to main src and test folders
