+++ packages/salesforcedx-vscode-soql/jest.config.js
@@ -1,7 +1,4 @@
 // eslint:disable-next-line:no-var-requires
 const baseConfig = require('../../config/jest.base.config');
 
-module.exports = {
-  ...baseConfig,
-  testMatch: [...baseConfig.testMatch, '<rootDir>/out/soql-common/**/*.test.js']
-};
+module.exports = Object.assign({}, baseConfig);
