+++ packages/salesforcedx-vscode-soql/src/editor/queryRunner.ts
@@ -8,8 +8,8 @@
 import type { Connection } from '@salesforce/core';
 import type { JsonMap } from '@salesforce/ts-types';
 import * as vscode from 'vscode';
-import * as soqlComments from '../../soql-common/soqlComments';
 import { nls } from '../messages';
+import * as soqlComments from '../soql-common/soqlComments';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 type QueryResult<T> = Awaited<ReturnType<Connection['query']>>;
