+++ packages/salesforcedx-vscode-soql/src/soql-common/soqlComments.ts
