+++ packages/salesforcedx-vscode-soql/tsconfig.json
@@ -7,7 +7,7 @@
     "noImplicitAny": false
   },
   "exclude": ["node_modules", ".vscode-test", "out", "./test/ui-test/resources/extensions", "dist"],
-  "include": ["src/**/*.ts", "test/**/*.ts", "soql-common/**/*.ts"],
+  "include": ["src/**/*.ts", "test/**/*.ts"],
   "references": [
     {
       "path": "../salesforcedx-sobjects-faux-generator"
